implement ProfileController

+++ src/main/java/com/project/ecommerce/service/implementation/StripeServiceImpl.java
@@ -57,6 +57,6 @@ public class StripeServiceImpl implements StripeService {
                         SessionCreateParams.LineItem.PriceData.ProductData.builder()
                                 .setName(stripeItemDto.getProductName())
                                 .build())
-                .setUnitAmount((long)(stripeItemDto.getPrice() * 100L)).build();
+                .setUnitAmount(stripeItemDto.getPrice() * 100L).build();
     }
 }
